requirements.txt
tensorflow>=2.6.0
numpy>=1.19.0
pandas>=1.3.0
Pillow>=8.0.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
pyyaml>=5.4.0